Add files via upload
Movie of temporal sensitivity at the earth's surface for a 5x5 m^2 source at the origin and a pointwise d_t B receiver at (20, 0 ,0). The grid has a size of 81x41x21 points and the time range spans 1e-7 to 1e-3 s. For the spline interpolation, 40000 equidistant values are used.